test-requirements.txt
ansible-core
pytest
mock
pytest-xdist
pytest-forked
